requirements.txt
google-cloud-vision>=3.4.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0 